requirements.txt
pandas
openai
anthropic
requests
nltk
python-dotenv
